--- src/Plugin/resource/DataProvider.php
@@ -1,10 +1,0 @@
-<?php
-
-/**
- * @file
- * Contains \Drupal\restful\Plugin\resource\DataProvider.
- */
-
-namespace Drupal\restful\Plugin\resource;
-
-class DataProvider implements DataProviderInterface {}
