+++ src/Plugin/resource/DataProvider/DataProvider.php
@@ -1,0 +1,10 @@
+<?php
+
+/**
+ * @file
+ * Contains \Drupal\restful\Plugin\resource\DataProvider\DataProvider.
+ */
+
+namespace Drupal\restful\Plugin\resource\DataProvider;
+
+class DataProvider implements DataProviderInterface {}
