+++ src/Plugin/resource/DataProvider/DataProviderInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+/**
+ * @file
+ * Contains \Drupal\restful\Plugin\resource\DataProvider\DataProviderInterface.
+ */
+
+namespace Drupal\restful\Plugin\resource\DataProvider;
+
+use Drupal\restful\Plugin\resource\CrudInterface;
+
+interface DataProviderInterface extends CrudInterface {}
