--- src/Plugin/resource/DataProviderInterface.php
@@ -1,10 +1,0 @@
-<?php
-
-/**
- * @file
- * Contains \Drupal\restful\Plugin\resource\DataProviderInterface.
- */
-
-namespace Drupal\restful\Plugin\resource;
-
-interface DataProviderInterface extends CrudInterface {}
