+++ src/Plugin/resource/Field/ResourceFieldEntity.php
@@ -2,10 +2,10 @@
 
 /**
  * @file
- * Contains \Drupal\restful\Plugin\resource\ResourceFieldEntity
+ * Contains \Drupal\restful\Plugin\resource\Field\ResourceFieldEntity
  */
 
-namespace Drupal\restful\Plugin\resource;
+namespace Drupal\restful\Plugin\resource\Field;
 
 use Drupal\restful\Exception\ServerConfigurationException;
