+++ src/Plugin/resource/Field/ResourceFieldInterface.php
@@ -2,10 +2,10 @@
 
 /**
  * @file
- * Contains \Drupal\restful\Plugin\resource\ResourceFieldInterface.
+ * Contains \Drupal\restful\Plugin\resource\Field\ResourceFieldInterface.
  */
 
-namespace Drupal\restful\Plugin\resource;
+namespace Drupal\restful\Plugin\resource\Field;
 
 use Drupal\restful\Exception\ServerConfigurationException;
